refactor: OAuth 인증페이지로 route하는 함수 개선
